Fix for crates using relative license-file paths in the Cargo.toml.
A license-file which didn't start with LICENSE was not handled correctly.

There are currently two ways which this crate uses to discover license
files.

The first method searches for paths starting with "LICENSE"
in the root of the current crate.
The second, uses a path specified by a crate author (in the Cargo.toml,
under the package table, a license-file can be specified).

The paths collected by the first method are absolute.
The path collected by the second however is relative to the crate.

The program however expects the paths to be absolute. For example,
in 'get_expression()', paths will be made relative by using the
'strip_prefix' method on the path. If the path is already relative,
as was the case before this commit, the program crashes.

Now perhaps the bug can also be fixed by making all paths relative,
however since I'm not well acquainted with the code base it is difficult
to make such a judgement, and as such this fix seemed safer to me.