Remove 'fix' functionality
This functionality was only added when there was the potential for
cargo-deny to supplant cargo-audit, which is no longer the case. This
functionality was quite a lot of code, minimally tested and increased
compile times (toml_edit), all for a feature that, at least
anecdotally, is not really useful as any moderately complex dependency
graph is unlikely to be patchable.