Update libgit2 (#422)
* Replace structopt with clap 3

* Remove rustsec git dependency

Rustsec has been really slow updating to a version with git 0.14 to
resolve a segfault, which also blocked cargo-deny. Instead, we now
handle all git fetching ourselves so we are no longer blocked, now or in
the future. This also changes it so that we use crates-index as a
dependency as previously we had included this functionality directly as
at the time crates-index didn't support bare repositories.

* Add metadata

* Remove unused dependencies

* Set MSRV to 1.60.0 due to cargo 0.61