Fix erroneously pruned crates (#481)
* Don't fail license gathering if expression is made

If _any_ license file failed the gather would fail, kind of defeating
the purpose of being able to have multiple license files. The failures
are now also added to the Ok if at least
some expression was able to be procured, and those warnings/errors are
now displayed in the diagnostic for the krate license. We also now
include all of the files that were involved in synthesizing the
expression for clarity

* Bump krates to 0.12.5

* Update CHANGELOG

* Tuple -> struct

* Update snapshots