Add trusted crate publishers
Similar to how we've done in a few of our other reviews, so we can focus on audits of more unknown crates and publishers.

Also excluded auditing for a few additional crates that are not used in practice (unsupported/unused targets).

Current audit state:

`Vetting Succeeded (136 fully audited, 26 partially audited, 107 exempted)`