Fix a mistake made in #531
I had initially patterned the new license checks on `OsiOnly` and `FsfOnly` before realizing that perhaps deny!() didn't make sense here, and I removed it (or so I thought). It seems I blundered and left one in--sorry.

I believe this deny!() shouldn't be here.

Apologies for the mistake...

(Also, thanks for catching that other mistake I made in #531 Jake)