Remove the misleading comment about RUSTSEC from deny.template.toml
Removes the comment from the template file. The comment might lead to people think there is no reason to use the `advisories/notice` option any more due to that comment.

The https://github.com/rustsec/advisory-db is updated regularly, so people may need to spend time thinking whether they need those or not instead of thinking that since 2019-12-17, there has been nothing, and so it is pointless now.