Allow for the dependency level checking for the bans of executables.
Changes the specification of the bans for executable checks so that
instead of just having the check as an on/off toggle, it allows to
specify the level of depth for a dependency within the dependency tree
(graph) from the root (the project being checked).

This allows to check for executables at a much more precise level.
Some projects may disallow for their direct dependencies to have
executables, but don't really care for the farthest dependencies.

A piece of thought for the future is to allow/disallow the test code
of the dependencies to have binaries. Some crates like libloading
legitimately can't be tested without having a binary to test against.