Fix issue with bans codes as args (#615)
An issue was introduced in [#605
](https://github.com/EmbarkStudios/cargo-deny/pull/605#discussion_r1501375569)
as I had commented out code while iterating on changes and forgot to add
back this case because it wasn't covered by any tests. I've added a
snapshot test to catch it in the future, as well as added the
`deprecated` code to the list of options.