Additional 1.85.0 fixup (#746)
This fixes an issue where crate sources were not being correctly
determined as crates.io when using cargo 1.85.0 due to the change in how
directory names are calculated, which would cause eg. workspace
dependency checks to fail.

The advisory databases no longer use tame_index to perform the directory
naming as it doesn't really make sense in the first place, so now it
just uses a similar method, but using xxhash instead which gives the
same properties as the cargo 1.85.0 change, namely that the hash is the
same for the same url regardless of the host platform
(endianness/pointer width/arch). It also now uses the last path
component as the start of the name which is a much nicer way to name the
directory.

This also updates to the rust-version to 1.85.0 and moves to edition
2024.