Transparently ignore semver trick (#773)
The [semver trick](https://github.com/dtolnay/semver-trick) has been
around for a while, but isn't widely used, however, recently
`webpki-roots` did use it for 0.26.11 -> 1.0.0, which means that every
crate with a dependency on it that banned duplicates needed to add a
skip. With this change, cargo-deny will detect if the only dependents of
a particular version are other versions of itself, skipping them
automatically.