Add a config option to suppress `unmatched-source` (#782)
Thank you for the nice project!
This is my first PR for this project.
Therefore if I have any mistakes for contribution, please let me know...

---

There are several reasons you might want this:

1. If the `allow` list represents the set sources for a project that
have gone through some external approval process, such as vetting it
with a legal department.
2. You're checking a single project in a workspace that shares its
`deny.toml` for all members, but not all members have identical
dependency sets.
3. You'd like to use `deny.toml` as part of a project template, and
configure it with some default set of sources you find acceptable
4. ... others, for sure...

As it is, this warning isn't a huge deal, but is annoying/unhelpful if
you don't care about it. I suspect that it's useful for catching typos
or keeping configuration tight (and so I think "warn" is the right
default for it), but there are enough reasons to want to turn it off
that it seems justified to me for it to be an option.

It was easy to add support for a config property which controls the lint
level for this check, so I just did that. I guess setting it to `deny`
could be desirable in some cases, although it seems a little dodgy to me
for various reasons... That said, I didn't see a reason to forbid that
sort of thing, and allowing it to be configured as a `LintLevel` seemed
more consistent.

### Additional Information

- This PR is for #781 
- This PR modification is very similar with #368 
- config name, `unused-allowed-source` is inspired by
`licenses.unused-allowed-licens` in #368