Add `markdown` output format to the list command
This format follows the existing 'human' option, but makes it easier to
render as a table in Markdown viewers. Both layouts are present:
 - Crate
 - License

The backtick (`) is used to make it easier to parse the long list of
elements in the second column. This design is partially inspired by the
cabal-plan license-report tool [1][2].

Sources:
[1] https://hackage.haskell.org/package/cabal-plan
[2] https://hackage.haskell.org/package/cabal-plan-0.7.6.1/src/example/cabal-plan.md